refactor: rename `nav` schematic to `navigation`
Our docs often describe the `nav` schematic as `navigation` schematics,
and it's not quite clear what the actual schematic name is. To avoid
this confusion, we just name it `navigation`, but keep support for
the old `nav` abbreviation.
