+++ src/material/schematics/ng-generate/navigation/schema.json
@@ -1,7 +1,7 @@
 {
   "$schema": "http://json-schema.org/schema",
-  "id": "SchematicsMaterialNav",
-  "title": "Material Nav Options Schema",
+  "id": "SchematicsMaterialNavigation",
+  "title": "Material Navigation Options Schema",
   "type": "object",
   "properties": {
     "path": {
