+++ src/material/schematics/ng-generate/navigation/schema.ts
